refactor: network/src/types-ambient.js
