+++ packages/network/src/types-ambient.js
